pyproject.toml
[project]
name = "fastapi-template"
version = "0.0.1"
description = ""
license = { text = "MIT" }
authors = [
    { name = "Doctor", email = "thirvondukr@gmail.com" },
]
requires-python = ">=3.13"
dependencies = [
    "aioinject>=1.4.3",
    "alembic>=1.16.3",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.0",
    "orjson>=3.10.18",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "result>=0.17.0",
    "sentry-sdk>=2.32.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "strawberry-graphql>=0.275.5",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0 ; sys_platform != 'win32'",
]


[dependency-groups]
lint = [
    "deptry>=0.23.0",
    "mypy>=1.16.1",
    "ruff>=0.12.2",
]
test = [
    "anyio>=4.6.2.post1",
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-archon>=0.0.6",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "sqlalchemy-pytest>=0.3.2.post1",
    "typeguard>=4.4.4",
]

[tool.uv]
default-groups = "all"


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["src/app/connectors/db/migrations/*"]
concurrency = ["greenlet"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "class .*\\(.*\\bProtocol\\b.*\\):",
    "@(?:typing\\.)?overload",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
skip_covered = true
skip_empty = true
precision = 2
fail_under = 100


[tool.deptry]
[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn", "uvloop", "asyncpg", "orjson"]

[tool.deptry.package_module_name_map]
uvloop = "uvloop"

[tool.mypy]
plugins = [
    "pydantic.mypy",
    "strawberry.ext.mypy_plugin",
]

strict = true
follow_imports = "normal"
ignore_missing_imports = false

allow_redefinition = false
disallow_any_explicit = false
ignore_errors = false
local_partial_types = true
no_implicit_optional = true
strict_optional = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true

pretty = true
show_column_numbers = true
show_error_codes = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.ruff]
src = ["src", "tests"]
[tool.ruff.lint]
fixable = [
    "F",
    "E",
    "W",
    "I",
    "COM",
    "UP",
    "RUF",
]
unfixable = [
    "F841", # Variable  is assigned to but never used
]
select = ["ALL"]
ignore = [
    "E501", # Line Length
    "D10", # Disable mandatory docstrings
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "TC001",
    "TC002",
    "TC003",
    "UP040", # Personally would prefer to not enable that as PyCharm still doesn't autocomplete this
    "COM812", # ruff format conflict
    "ISC001", # ruff format conflict
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"src/app/connectors/db/migrations/*" = ["INP001"]
"src/app/connectors/db/migrations/versions/*" = ["D400", "D415", "N999"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.mccabe]
max-complexity = 6

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"starlette.status".msg = "Use http.HTTPStatus instead."
"fastapi.status".msg = "Use http.HTTPStatus instead."
"litestar.status_codes".msg = "Use http.HTTPStatus instead."
"httpx.codes".msg = "Use http.HTTPStatus instead."
